refactor(sharing-links): extract user avatar components into own file
Move UserAvatar, UserAvatarWithLabel and their helpers out of the cell
renderer into a dedicated user-avatar module, so both the list cell
renderer and the sharing-links dialog reuse them.
