+++ src/components/dialogs/sharing-links/sharing-links-dialog.tsx
@@ -5,7 +5,6 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 import {
-    Box,
     Button,
     Dialog,
     DialogActions,
@@ -21,20 +20,10 @@ import { FormattedMessage, useIntl } from 'react-intl';
 import { useState } from 'react';
 import type { UUID } from 'node:crypto';
 import { type ElementAttributes, ElementType } from '@gridsuite/commons-ui';
-import { UserAvatar } from '../../utils/renderers/user-cell-renderer';
+import { UserAvatarWithLabel } from '../../utils/renderers/user-avatar';
 import { DateCellRenderer } from '../../utils/renderers/date-cell-renderer';
 import { getElementTypeTranslation } from '../../utils/translation-utils';
 import PathBreadcrumbs from './path-breadcrumbs';
-
-/** A user name preceded by its GridExplore list avatar, in a compact size. */
-function UserAvatarWithLabel({ label }: Readonly<{ label: string }>) {
-    return (
-        <Box sx={{ display: 'inline-flex', alignItems: 'center', gap: 1 }}>
-            <UserAvatar label={label} size="small" />
-            {label}
-        </Box>
-    );
-}
 
 /**
  * One element using the inspected shared element (e.g. a study referencing a shared
